Merged in some minor bug fixes, enhancements, and API changes.

+++ src/gfx_adapter.c
@@ -41,7 +41,7 @@ static void convert_uv_to_atlas( float *atlas_uv_out, short tc[] )
 
 static void process_display_list( void *dl )
 {
-    int64_t *ptr = (int64_t *)dl;
+    intptr_t *ptr = (intptr_t *)dl;
     Vtx *vdata = NULL;
 
     for( ;; )
@@ -50,23 +50,23 @@ static void process_display_list( void *dl )
         {
             case GFXCMD_VertexData: 
             {
-                UNUSED int64_t v = *ptr++;
-                UNUSED int64_t n = *ptr++;
-                UNUSED int64_t v0 = *ptr++;
+                UNUSED intptr_t v = *ptr++;
+                UNUSED intptr_t n = *ptr++;
+                UNUSED intptr_t v0 = *ptr++;
                 vdata = (Vtx*)v;
                 break;
             }
 
             case GFXCMD_Triangle:
             {
-                int64_t v00 = *ptr++;
-                int64_t v01 = *ptr++;
-                int64_t v02 = *ptr++;
-                UNUSED int64_t flag0 = *ptr++;
+                intptr_t v00 = *ptr++;
+                intptr_t v01 = *ptr++;
+                intptr_t v02 = *ptr++;
+                UNUSED intptr_t flag0 = *ptr++;
 
-                short x0 = vdata[v00].v.ob[0], y0 = vdata[v00].v.ob[1], z0 = vdata[v00].v.ob[2];
-                short x1 = vdata[v01].v.ob[0], y1 = vdata[v01].v.ob[1], z1 = vdata[v01].v.ob[2];
-                short x2 = vdata[v02].v.ob[0], y2 = vdata[v02].v.ob[1], z2 = vdata[v02].v.ob[2];
+                float x0 = vdata[v00].v.ob[0], y0 = vdata[v00].v.ob[1], z0 = vdata[v00].v.ob[2];
+                float x1 = vdata[v01].v.ob[0], y1 = vdata[v01].v.ob[1], z1 = vdata[v01].v.ob[2];
+                float x2 = vdata[v02].v.ob[0], y2 = vdata[v02].v.ob[1], z2 = vdata[v02].v.ob[2];
                 Vec3f p0 = { (float)x0, (float)y0, (float)z0 };
                 Vec3f p1 = { (float)x1, (float)y1, (float)z1 };
                 Vec3f p2 = { (float)x2, (float)y2, (float)z2 };
@@ -129,8 +129,8 @@ static void process_display_list( void *dl )
 
             case GFXCMD_Light:
             {
-                int64_t l = *ptr++;
-                int64_t n = *ptr++;
+                intptr_t l = *ptr++;
+                intptr_t n = *ptr++;
 
                 if( n == 1 )
                 {
@@ -145,9 +145,9 @@ static void process_display_list( void *dl )
 
             case GFXCMD_Texture:
             {
-                int64_t s = *ptr++;
-                int64_t t = *ptr++;
-                int64_t on = *ptr++;
+                intptr_t s = *ptr++;
+                intptr_t t = *ptr++;
+                intptr_t on = *ptr++;
 
                 s_scaleS = (uint16_t)s;
                 s_scaleT = (uint16_t)t;
@@ -158,7 +158,7 @@ static void process_display_list( void *dl )
 
             case GFXCMD_SetTextureImage:
             {
-                int64_t i = *ptr++;
+                intptr_t i = *ptr++;
 
                 s_textureIndex = (int)i;
                 s_texWidth = mario_tex_widths[s_textureIndex];
@@ -169,8 +169,8 @@ static void process_display_list( void *dl )
 
             case GFXCMD_SetTileSize:
             {
-                int64_t uls = *ptr++;
-                int64_t ult = *ptr++;
+                intptr_t uls = *ptr++;
+                intptr_t ult = *ptr++;
                 ptr++; // lrs
                 ptr++; // lrt
 
@@ -182,7 +182,7 @@ static void process_display_list( void *dl )
 
             case GFXCMD_SubDisplayList:
             {
-                int64_t dl = *ptr++;
+                intptr_t dl = *ptr++;
                 process_display_list( (void*)dl );
                 break;
             }
